few things

+++ GUI_Combine/table2robot.m
@@ -1,3 +1,5 @@
+% convert pixel coordinate from TABLE camera to robot coordinate 
+
 function [Xr,Yr] = table2robot(Xt,Yt)
 R = [0 1 ; 1 0];
 S = [0.65 0 ; 0 0.65];
@@ -5,6 +7,6 @@ T = [77.5 ; -520];
 TCoor = [Xt ; Yt];
 
 RCoor = R*S*TCoor + T;
-Xr = int32(RCoor(1));
-Yr = int32(RCoor(2));
+Xr = (RCoor(1));
+Yr = (RCoor(2));
 return
